Refactor modal
